refactor(dash): improve validatorMap
